move VMaaS CVE sync procedure to vmaas-sync service
