+++ common/utils.py
@@ -1,0 +1,35 @@
+"""
+Various utility functions.
+"""
+
+import os
+
+from prometheus_client import Counter
+import requests
+
+from common.logging import get_logger
+
+VMAAS_REQUEST_RETRIES = int(os.getenv("VMAAS_REQUEST_RETRIES", "3"))
+
+LOGGER = get_logger(__name__)
+
+# prometheus probes
+# counts
+VMAAS_RETURN_ERR = Counter('ve_evaluator_vmaas_return_errors', 'Number of non-200 RCs from VMaaS')
+VMAAS_CNX_ERR = Counter('ve_evaluator_vmaas_cnx_errors', 'Number of connection-errors from VMaaS')
+
+def vmaas_post_request(endpoint, data_json):
+    """Sends request to VMAAS"""
+    for _ in range(VMAAS_REQUEST_RETRIES):
+        try:
+            response = requests.post(endpoint, json=data_json)
+            if response.status_code == 200:
+                return response.json()
+            VMAAS_RETURN_ERR.inc()
+            LOGGER.error("Error during request to VMaaS endpoint %s: HTTP %s, %s",
+                         endpoint, response.status_code, response.text)
+            LOGGER.debug("JSON: %s", str(data_json))
+        except ConnectionError:
+            VMAAS_CNX_ERR.inc()
+            LOGGER.exception("Error calling VMAAS: ")
+    return None
