+++ evaluator/evaluator.py
@@ -5,21 +5,19 @@ vulnerability-engine evaluator
 
 import os
 import json
-import requests
 from prometheus_client import Counter, Histogram, start_http_server
 from psycopg2.extras import execute_values
 
 from common import mqueue
 from common.database_handler import DatabaseHandler, init_db
 from common.logging import init_logging, get_logger
+from common.utils import vmaas_post_request
 
 LOGGER = get_logger(__name__)
 
 VMAAS_HOST = os.getenv('VMAAS_HOST', 'http://vmaas-webapp-1.vmaas-ci.svc:8080') # pylint: disable=invalid-name
-VMAAS_REQUEST_RETRIES = int(os.getenv("VMAAS_REQUEST_RETRIES", "3"))
 vmaas_updates_endpoint = "%s/api/v1/updates" % VMAAS_HOST # pylint: disable=invalid-name
 vmaas_errata_endpoint = "%s/api/v1/errata" % VMAAS_HOST # pylint: disable=invalid-name
-vmaas_cves_endpoint = "%s/api/v1/cves" % VMAAS_HOST # pylint: disable=invalid-name
 
 kafka_evaluator_topic = os.getenv('EVALUATOR_TOPIC', 'vulnerability.evaluator.requests') # pylint: disable=invalid-name
 prometheus_port = os.getenv('PROMETHEUS_PORT', '8085') # pylint: disable=invalid-name
@@ -28,8 +26,6 @@ prometheus_port = os.getenv('PROMETHEUS_PORT', '8085') # pylint: disable=invalid
 # times
 VMAAS_EVAL_TIME = Histogram('evaluator_vmaas_evaluation_seconds', 'Time spent checking a system for vmaas hits')
 # counts
-VMAAS_RETURN_ERR = Counter('ve_evaluator_vmaas_return_errors', 'Number of non-200 RCs from VMaaS')
-VMAAS_CNX_ERR = Counter('ve_evaluator_vmaas_cnx_errors', 'Number of connection-errors from VMaaS')
 VMAAS_COUNT = Counter('ve_evaluator_vmaas_calls', 'Number of VMaaS-evaluations attempted')
 INV_ID_NOT_FOUND = Counter('ve_evaluator_inventory_not_found', 'Number of times inventory-id not in SystemPlatform')
 UNKNOWN_MSG = Counter('ve_evaluator_unknown_msg', 'Number of unrecognized messages delivered from queue')
@@ -53,16 +49,6 @@ class QueueEvaluator:
         # get DB connection
         init_db()
         self.conn = DatabaseHandler.get_connection()
-        self.impact_id_map = self._prepare_impact_id_map()
-
-    def _prepare_impact_id_map(self):
-        impact_id_map = {}
-        cur = self.conn.cursor()
-        cur.execute("select name, id from cve_impact")
-        for impact_name, impact_id in cur.fetchall():
-            impact_id_map[impact_name] = impact_id
-        cur.close()
-        return impact_id_map
 
     def _load_cves_for_inventory_id(self, inventory_id):
         system_cves_map = {}
@@ -77,6 +63,9 @@ class QueueEvaluator:
             system_cves_map[cve] = sys_vuln
         cur.close()
         return system_cves_map
+
+    def _register_missing_cves(self, missing_cves):
+        pass
 
     def _store_new_cves(self, inventory_id, new_cves):
         system_cves_id_map = {}
@@ -94,21 +83,15 @@ class QueueEvaluator:
         cve_system_list = [(inventory_id, cve) for cve in new_cves]
 
         # Insert new CVEs into db
-        if cve_metadata_missing:
-            cves_synced = self._sync_cve_md(cve_list=cve_metadata_missing)
-        else:
-            cves_synced = True
+        self._register_missing_cves(cve_metadata_missing)
 
-        if cves_synced:
-            execute_values(cur, "insert into system_vulnerabilities \
-                           (inventory_id, cve) values %s \
-                           returning cve, id",
-                           cve_system_list, page_size=len(cve_system_list))
+        execute_values(cur, "insert into system_vulnerabilities \
+                       (inventory_id, cve) values %s \
+                       returning cve, id",
+                       cve_system_list, page_size=len(cve_system_list))
+        for cve, sys_vuln_id in cur.fetchall():
+            system_cves_id_map[cve] = sys_vuln_id
 
-            for cve, sys_vuln_id in cur.fetchall():
-                system_cves_id_map[cve] = sys_vuln_id
-        else:
-            LOGGER.error("Unable to sync CVEs from VMAAS for system: %s", inventory_id)
         cur.close()
         self.conn.commit()
         return system_cves_id_map
@@ -198,73 +181,6 @@ class QueueEvaluator:
         cur.close()
         self.conn.commit()
 
-    def _sync_cve_md(self, page_size=5000, cve_list=None):
-        LOGGER.info('Syncing CVE metadata')
-        cur = self.conn.cursor()
-        cur.execute('select cve from cve_metadata')
-        cves_in_db = []
-        for cve_tuple in cur.fetchall():
-            cves_in_db.append(cve_tuple[0])
-        # Fetch all CVEs if cve_list is not specified
-        if not cve_list:
-            cve_list = [".*"]
-        success = True
-        page = 1
-        while True:
-            cve_request = {'cve_list' : cve_list, 'page_size' : page_size, 'page': page, 'rh_only' : True}
-            LOGGER.info('Downloading CVE metadata (page: %s, page_size: %s)', page, page_size)
-            r_json = vmaas_post_request(vmaas_cves_endpoint, cve_request)
-            if r_json is None:
-                success = False
-                break
-            LOGGER.info('Importing CVE metadata (page: %s, page_size: %s, pages: %s)',
-                        page, page_size, r_json['pages'])
-            cves = r_json['cve_list']
-            to_insert = []
-            to_update = []
-            for cve in cves:
-                description = cves[cve]['description']
-                impact_id = self.impact_id_map[cves[cve]['impact']]
-                public_date = cves[cve]['public_date'] or None
-                modified_date = cves[cve]['modified_date'] or None
-                cvss3_score = float(cves[cve]['cvss3_score']) if cves[cve].get('cvss3_score') else None
-                cvss3_metrics = cves[cve].get('cvss3_metrics')
-                cvss2_score = float(cves[cve]['cvss2_score']) if cves[cve].get('cvss2_score') else None
-                cvss2_metrics = cves[cve].get('cvss2_metrics')
-                row = (cve, description, impact_id, public_date, modified_date, cvss3_score, cvss3_metrics,
-                       cvss2_score, cvss2_metrics)
-                if cve not in cves_in_db:
-                    to_insert.append(row)
-                else:
-                    to_update.append(row)
-            if to_insert:
-                execute_values(cur, """insert into cve_metadata
-                                    (cve, description, impact_id, public_date, modified_date,
-                                    cvss3_score, cvss3_metrics, cvss2_score, cvss2_metrics)
-                                    values %s""", to_insert, page_size=len(to_insert))
-            if to_update:
-                execute_values(cur, """update cve_metadata set description = data.description,
-                               impact_id = data.impact_id,
-                               public_date = cast(data.public_date as timestamp with time zone),
-                               modified_date = cast(data.modified_date as timestamp with time zone),
-                               cvss3_score = cast(data.cvss3_score as numeric),
-                               cvss3_metrics = data.cvss3_metrics,
-                               cvss2_score = cast(data.cvss2_score as numeric),
-                               cvss2_metrics = data.cvss2_metrics
-                               from (values %s) as data (cve, description, impact_id, public_date, modified_date,
-                               cvss3_score, cvss3_metrics, cvss2_score, cvss2_metrics)
-                               where cve_metadata.cve = data.cve""",
-                               to_update, page_size=len(to_update))
-            LOGGER.info('Finished importing CVE metadata (page: %s, page_size: %s, pages: %s)',
-                        page, page_size, r_json['pages'])
-            if page >= r_json['pages']:
-                break
-            page += 1
-        cur.close()
-        self.conn.commit()
-        LOGGER.info('Finished syncing CVE metadata')
-        return success
-
     @staticmethod
     def _vmaas_request_cves(updates_json):
         """Make VMaaS request for updates, then cves returning them"""
@@ -339,11 +255,6 @@ class QueueEvaluator:
 
         # pylint: disable=too-many-branches
         cur = self.conn.cursor()
-        cur.execute('select count(*) from cve_metadata')
-        cve_md_count = int(cur.fetchone()[0])
-        if cve_md_count == 0:
-            LOGGER.info('no cve_metadata found, attempting to pull them from vmaas')
-            self._sync_cve_md()
 
         def process_message(message):
             """Message procession logic"""
@@ -352,15 +263,7 @@ class QueueEvaluator:
                 if 'type' not in msg_dict:
                     LOGGER.error("Received message is missing type field: %s", msg_dict)
                     return
-                if msg_dict['type'] == 'vmaas_updates':
-                    LOGGER.info("Received message type: %s", msg_dict['type'])
-                    LOGGER.info("VMaaS metadata changed, need to re-evaluate all systems")
-                    self._sync_cve_md()
-                    cur.execute("select inventory_id from system_platform")
-                    # reevaluate updates for every system in the DB
-                    for inventory_id, in cur.fetchall():
-                        self.producer.send({"type": "re-evaluate_system", "system_id": inventory_id})
-                elif msg_dict['type'] in ['upload_new_file', 're-evaluate_system']:
+                if msg_dict['type'] in ['upload_new_file', 're-evaluate_system']:
                     LOGGER.info("Received message type: %s", msg_dict['type'])
                     cur.execute("""select inventory_id, vmaas_json, rh_account,
                                           opt_out, satellite_managed from system_platform
@@ -386,23 +289,6 @@ def run_evaluator():
     """Sets up and runs the evaluator"""
     evaluator = QueueEvaluator([kafka_evaluator_topic])
     evaluator.run()
-
-
-def vmaas_post_request(endpoint, data_json):
-    """Sends request to VMAAS"""
-    for _ in range(VMAAS_REQUEST_RETRIES):
-        try:
-            response = requests.post(endpoint, json=data_json)
-            if response.status_code == 200:
-                return response.json()
-            VMAAS_RETURN_ERR.inc()
-            LOGGER.error("Error during request to VMaaS endpoint %s: HTTP %s, %s",
-                         endpoint, response.status_code, response.text)
-            LOGGER.debug("JSON: %s", str(data_json))
-        except ConnectionError:
-            VMAAS_CNX_ERR.inc()
-            LOGGER.exception("Error calling VMAAS: ")
-    return None
 
 
 def main():
