+++ tests/evaluator_tests/test_evaluator.py
@@ -40,7 +40,6 @@ class QueueEvaluatorTest(QueueEvaluator):
     def __init__(self, pg_db_conn):
         """Override __init__ to not init database, DB is initialized in pg_db_conn fixture."""
         self.conn = pg_db_conn
-        self.impact_id_map = self._prepare_impact_id_map()
 
 
 class TestEvaluator:
